feat: Setup CLI for project initialization
